Fix file-system integration test to use clear write command
The test was failing because gemini-2.5-flash was being cautious about
using the replace tool on an empty file. Changed the prompt from
'edit test.txt to have a hello world message' to 'write "Hello World\!" to test.txt'
to make it clear that the write_file tool should be used instead.